requirements.txt
playwright
pytest
pytest-bdd
allure-pytest
python-dotenv
